Tests

+++ KillerApp/Models/Beoordeling.cs
@@ -10,7 +10,7 @@ namespace KillerApp.Models
         public int Id { get; set; }
         public string Titel { get; set; }
         public string Omschrijving { get; set; }
-        public double Cijfer { get; set; }
+        public decimal Cijfer { get; set; }
 
         public Product Product { get; set; }
     }
